Experiments for thesis

+++ Ender/Rule.py
@@ -1,3 +1,5 @@
+import numpy
+
 class Rule():
     def __init__(self):
         self.decision = None
@@ -25,8 +27,8 @@ class Rule():
         self.conditions.append(condition)
         self.attribute_names.append(attribute_name)
 
-    def classify_instance(self, x, prune=False): # TODO delete prune
+    def classify_instance(self, x, prune=False):
         for condition in self.conditions:
             if not condition[1] <= x[condition[0]] <= condition[2]:
-                return 0 if type(self.decision) is int else [0 for _ in range(len(self.decision))]
+                return 0 if type(self.decision) is numpy.float64 else [0 for _ in range(len(self.decision))]
         return self.decision
